Improve UX of docs drawer menu (#194)
This PR introduces a `StickyScroll` module that tracks the scroll position of an element and retains it's position when navigating to a different page. The script also checks that the active menu item is visible and adjusts the scroll position if it's not. Also introduced in this work is an updated drawer menu style to look more like tabs.