Fix modal hidden focusable bug (#280)
This PR extracts the focusable array build into it's own method and includes a focus test. This also created the need to fire open and close sequentially so there was a refactor introducing promises so that an open call only happens after a close is complete. The impacts how opening modals from inside other modals works. Because of this refactor, it removes the need for callbacks since you can now wait for the promises to resolve. Lastly, a "working" check is done on click so that it's no longer possible to trigger a new modal open if one is in the middle of transitioning.