Update package entries and bundle output (#312)
This PR removes the CommonJS and IIFE bundles in favor of the UMD pattern for `scripts.js` and `scripts.min.js`. The new `main` package field now points to the UMD `scripts.js` for better compatibility with other bundlers via NPM and `node_modules`. Also included is the `modules` field which points to the ES6 entry for components.