Update spacing (#335)
This refactors the `spacing()` mixin to only apply margin top using the `> * + *` selector and updates all spacing map values to use `em`s instead of `rem`s. In addition to these refactors, this PR also adds the ability to pass a map key to `spacing()` that checks if a value exists within the `spacing-map` variable.