Change the order or args in moveElement module (#339)
This PR refactors the the argument order for both the `moveElement` module and `modalMove` api in the modal component. The new order is as follows: target (the elements being moved), type (the type of move, before, after etc), reference (the reference element to apply the move on). The new order now reads more clearly as "move element '.selector' 'after' '.reference'.