Audit and refactor section component (#434)
This PR audits and refactors the section component to be more robust and have a more clear purpose. This work also includes adding website and README documentation. The primary changes include adding similar container styles that existed in the now deprecated `container` component. Another non-breaking change includes centering the container within the component regardless of the set height or min-height.