Audit and refactor menu component (#607)
This PR is a complete audit and partial refactor of the menu component. It removes unneeded modifiers and themes while providing more intuitive default styles as well as a refactor of variable names and improvements to state styles. This PR also includes full documentation of the component. The following are breaking changes:

- Menu now stacks as the default styles and removes the `menu_stack_[value]` modifier.
- Adds the new `menu_inline_[value]` modifier for making a menu inline (horizontal styles).
- Adds new `menu_size_[value]` modifier and matches sizing scale to other form-control components.
- Refactors, renames and deprecates a number of menu variables.