Refactor dropdown component, rename to popover and implement popper js (#614)
This PR introduces the new popover component and deprecates dropdown. Popover is a robust container component that is initially hidden but can be revealed using a click or hover event. It's dynamically positioned using the [Popper positioning engine](https://popper.js.org/). The previous dropdown menu styles have also been deprecated in favor of a more compositional use of the menu component.