Further refactor Popover JS module and improve a11y (#838)
This PR refactors some of the core architecture of the popover JS module and adds some handy new features along with some accessibility enhancements. Popovers now require `aria-controls="unique-id"` and `id="unique-id"` attributes for triggers and targets respectively. This removes the need to add additional data attributes for popovers and their triggers.

Another big addition in this PR is the `Collection` class which `Popover` extends. This allows for a re-usable pattern to build and interface with collections. Lays the groundwork in collections for modal and drawer modules.

**Other changes include:**

- `open()` and `close()` methods now take popover IDs.
- `register()` and `deregister()` methods now take popover ID or either popover elements.
- Deprecate `closeAll()` in favor of calling `close()` without a parameter to close all popovers.
- Each popover object now has built in methods: `open()`, `close()` and `deregister()`.
- Deregister has been refactored to ensure event listeners and object properties are removed.
- New `get()` method for retrieving collection entries by any popover object property (defaults to `id`).
- Deprecate the need for data attributes to pass override settings in favor of CSS variables.