Remove the use of deleteProperty in localStore module (#893)
Proxy is supported by all major browsers, but in cases where a proxy polyfill is needed we should only use supported traps. This PR removes the use of `deleteProperty` and instead uses `set` to determine if a property should be deleted based on if it is being set to `undefined`.