Refactor modal screen styles and how they are applied (#900)
This PR refactors the modal screen styles to instead be applied to a pseudo element allowing us to set any color property for the background and controlling the transparency of the background using the separate `opacity` property. This PR also renames the related variables to prefix `screen-` for clarity and consistency.