Replace microbundle with Vite (#1308)
This PR replaces [microbundle](https://github.com/developit/microbundle) with [vite](https://github.com/vitejs/vite) as the module bundler and starts the groundwork for using vite `index.html` as a demo and testing tool. It also cleans up and standardizes the output naming convention for both scripts and styles. The following notable changes are included in this PR:

- Output files in `dev` and `dist` are now named `index.[ext]`, no more `scripts` and `styles` naming convention.
- Nodemon has been removed in favor of native sass watch.
- Added base `index.html` files to packages that utilize vite. These currently just log the output of `mjs` and `cjs` files.
- JavaScript format is now set to `type module` and removed the two extra builds that were created in the previous microbundle usage. Now only `mjs` and `umd` bundle formats are output.