Add back lerna-changelog and remove changesets (#1344)
This PR adds back `lerna-changelog` and removes the use of `changesets`. The reason for this is I was not a fan of the many tags and separate changelog files that changesets creates. I much better like the PR driven changelog that lerna-changelog creates and that it creates a single file where this information is stored.