Add the ability to set a custom screen for modals and modal drawers (#1365)
This PR adds the ability to set a custom screen element for closing modals. This by default is the root modal and modal drawer elements. It also updates the handlers appropriately and also adds a `required` property on the modal collection entries so it's easier to access that data point using `entry.required`.