Introduction of indeterminate state for better handling of inline state. (#1374)
This PR introduces the concept of "indeterminate" drawer state. This state allows users to better control the state of a drawer using CSS and not so heavily rely on JavaScript controlled classes for state management. To better track the inline drawer state, a new property is added to the entry object `entry.inlineState` and is restored when switching from modal to inline modes where `entry.state` will continue to hold the current state of the drawer.

The following state logic has been applied:
- Modal drawers are always closed as their initial state and when switching from inline to modal. If a modal needs to be opened on load or by other events, that should be done programmatically using the appropriate APIs.
- On initial registration: inline drawer state is determined in this order of priority:
  - Store state
  - Class state (`is-opened`, `is-closed`)
  - Falls back to indeterminate
- On switch from modal to inline: inline drawer state is determined in this order of priority:
  - Store state
  - inlineState var
  - Falls back to indeterminate