fix: add background and foreground theme to popover and dialog components (#1433)
Popovers and dialogs did not look very well on dark mode and that's because their foreground and backgrounds were not setup to use the new theme module system. This PR applies the background and foreground theme so that the content of popovers and modals are still readable in dark mode.