Add page navigation to docs (#1465)
This PR adds the `CollectionPagi` component which allows navigating forward and back between "previous" and "next" pages in a collection. This navigation appears at the bottom of the page. As part of this work `CollectionNavi` has been refactored along with how collections are queried in the page and collection index templates. There are also two new slots (`append` and `prepend`) in the base layout for displaying content above and/or below the primary content of the page.