Add "Suggest changes to this page" link to bottom of content (#1466)
This PR adds the a new "Suggest changes to this page" link at the bottom of content pages allowing users to make suggestions for changes on GitHub for documentation pages. This is rendered using the updated `append` layout slot and is either displayed by itself or above the page navigation component.