Update code example markup and styles (#1484)
This PR updates the markup and styles for the `CodeExample` component (previously named `Demo`) to match the new design. We also update the `CodeBlock` styles to both be more robust and compatible with being nested inside a code example. Code examples should always contain a code block and optionally the rendered output. They also display a "toolbar" with the language the code is written in and the copy code button.