Improve customization documentation (#1537)
This PR introduces a number of improvements to page documentation and applies them to the `button` components documentation page as the initial trial run. It establishes a way to consistently markup the structure of the page allowing for the most consistent and logic headings hierarchy as well as a new component for rendering the contents of files as well as mixins and functions. This also allows those content blocks to be toggled to help not overwhelm users with content as well as making it easier to scan what the subjects of the blocks are.

There are also some more minor improvements to a few modules, such as:

- Moved the copy button to appear after the code block and also made the code block focusable.
- Allow disable of anchor treatment on headings using the `no-anchor` class.
- New function for making a string safe in HTML attributes such as classes or IDs.
- Buttons now use the new theme module for outputting themes set on the button component.
- Removed redundant breakpoint SCSS vars.
- Fixed import order of root in notice component.
- Made content of rowed tables flush to table edges.
- Added utility `flex-items-fill` for allowing elements inside a flex container to fill the total width available.