Fix Astro hydration mismatch and remove Vue from docs (#1780)
With a recent Vue update (3.4), I was experiencing a lot of hydration mismatch warnings. This was mainly coming from the `popover` component used in `Pre`, `CodeBlock` and `ThemeSwitcherPopover`. One solution was to change the astro directive from `client:load` to `client:only="vue"` but this caused those components to "flash" into render on page load.

The solution that I decided to use in this PR was to convert all the vue components to astro components. This also allowed me to remove Vue as well as some typescript configs and dependencies.