Package documentation state (#1820)
This PR introduces a way to better track the state of documentation pages by displaying status icons in the package navigation items. This update is temporary as I work through updating the documentation format and content of components, but it should help keep things organized. In addition to the status indicator icon, there are also some general improvements made to the documentation:

- Inline drawer navigation now scrolls to active item.
- Applied new documentation format to core, utility and base packages.
- Improved the order that package items are listed in drawer navication.
- Created module for fetching the package entries (`getPackages.js`).
- Added more padding for "On this page" menu to properly clear the footer.