Refactor cssVar core function to not require the double dash (#1861)
This PR refactors the `cssVar()` function so that it no longer requires the passed property name to include the `--`. If the property name is passed without double dashes, it will be prefixed using `--vb-prefix` as well as have the `--` added. This allows passing a custom CSS variable property name that won't be modified as well a one that will get automatically prefixed. Here's an example of this logic:

- Passing `--background` will return the value of `--background`.
- Passing `background` will return the value of `--vb-background`.
- Passing `background` with no prefix variable set will return the value of `--background`.

The `prefix` parameter has been removed from `cssVar()` since this new logic makes that option redundant. The value of `vb-` previously came from the custom property `--vrembem-prefix`. This has been renamed to `--vb-prefix`, its value can be set in the core module `prefix.$variable`.

Fixes #1860