Improve error messages on modal and drawer event handlers (#1863)
This PR directly improves the error messages of Modal and Drawer event handlers by throwing an error if the element that is queried does not exist on the page. This is done by using `getModal` and `getDrawer` helper functions instead of the `this.get` which only returns `undefined` if entry does not exist in the collection.