Add css function fallback parameter and apply to grid gap use case (#1865)
This PR adds a new parameter option to the `core.css` function which allows setting a fallback CSS custom property. This is then applied to the grid component and the custom gap properties. This allows setting either a `grid-gap` value or a more specific `grid-gap-x`/`grid-gap-y` property instead of always having to set both the `-x` and `-y` variants.