Add the ability to reset code examples (#1867)
This PR introduces the `resettable` module to the documentation code examples. In cases where the code example output is demonstrating some JavaScript that changes the example state or HTML, we've added a "reset" button that restores the example to it's original state. This can be enabled as a `CodeExample` prop `resettable`. This update also:

- Removes the use of a value on boolean props.
- Puts `dismissible` JavaScript into its own module.
- Refactors `dismissible` to attach event listener to document instead of the HTML button.