Migrate the input component to custom property methods (#1877)
This PR migrates the input component to the new custom property (CSS variables) methodology. This applies new custom properties and properly applies the new BEM and CSS mixins/functions. The docs have also been updated to reflect these changes.

This PR also includes a refactor of the `core.css` function so that it can now take any number of property names as a list and they will be treated as fallbacks in the order they are provided.

Additional docs changes include:

- Made non-tab code examples resettable. 
- Applied a new status on form control components for future updates.
- Updated case on variants label.