Refactor core into a more defined library (#1882)
This PR is a complete refactor of the core Sass structure to make the package more of a library. It removes the use of general function, mixin and variable file in favor of a library and variables directory with separate files for things that have specific purposes or groups. It also creates a utilities directory for functions and mixins that are intended for internal use.

With the new variables directory, core variables are also all output as CSS custom properties. These outputs can be found in the `root` directory by the `palette.scss`, `theme.scss` and `variables.scss` files.

Lastly, a number of Sass functions and mixins were removed to help reduce unnecessary complexity within component styles and the core library.

**Package changes:**
- Vrembem: Moved root import to the top of vrembem manifest. Custom properties should now be defined first before they are used or overridden.
- Core: Added new debug-map helper mixin.
- Button: Removed some complexity around how the loading state is customized.
- All: Applied CSS variables to all components that were using the SCSS version of a core variable.
- All: Removed the use of `theme.$root-selector` for custom properties that are not intended to be used as a theme. These are now output under the `:root` selector alone and should not be effected by theme classes.

**Deprecated:**
- `@mixin gap()`
- `@mixin ellipsis()`
- `@function encodecolor()`
- `@function replace()`
- `@function set-border-color()`
- `@function theme.val()`

**Documentation changes:**
- Apply css mixin and function to doc styles.
- Fixed code-example overflow styles.
- Use gap utility on DocBlock to prevent over specification rule.