Add CSS module for managing custom properties (#1883)
This PR introduces the new CSS module for managing custom properties. The CSS module replaces the use of the css function and mixin from the core library which this PR also deprecates. CSS module can be imported and used using the new entry point:

```scss
@use "@vrembem/core/css";
@include css.log();
```

It can also be used with a core import where all its API is prefixed with its module name (same as palette and theme):

```scss
@use "@vrembem/core";
@include core.css-log();
```

**Deprecated:**
- `@function core.css()`: replaced with `css.get()`.
- `@mixin core.css()`: replaced with `css.set()` and `css.output()`.

**Additional changes:**
- Creates the core `modules` directory and updates export entry paths. This is where `css`, `palette` and `theme` now reside.
- Fixing broken search and replace of core.css usage.
- Replaces the usages of `core.$[var]` in favor of custom properties `css.get("[prop]")`.