Update github workflows (#1909)
This PR updates the existing build and deploy workflows to reduce redundancy by only running these flows under more specific conditions. Deploy should now only run when pushes are made to the `main` branch. Build is run on all pull requests as well as when a push is made to the `next` branch.