Add usage flags and conditional output (#1914)
This PR updates the `usage` component to now only output a custom property once. This can be circumvented by using the `"override"` action which is tracked as its own property in the meta map. 

The concept of "flags" has also been added to the meta map where a boolean flag can be set to alter a custom properties behavior. The first use-case is the "required" flag which allows a custom property to be output even if it was never referenced. This is useful for properties that are primarily used by JavaScript.