Improve package exports for better entry paths (#1918)
This PR improves the entry packs by creating better values in `package.exports`. There are two primary ways to consume a Vrembem component and two unique exceptions. A component that only offers CSS and SCSS modules, exports are now set as follows:

```js
"exports": {
  ".": {
    "sass": "./index.scss",
    "style": "./dist/index.css"
  },
  "./dev/*": "./dev/*",
  "./dist/*": "./dist/*",
  "./*": "./src/*"
},
```

Modules that also include JavaScript entries such as `drawer`, `modal` and `popover` now look like this:

```js
"exports": {
  ".": {
    "import": "./index.js",
    "require": "./dist/index.umd.cjs",
    "sass": "./index.scss",
    "style": "./dist/index.css"
  },
  "./dev/*": "./dev/*",
  "./dist/*": "./dist/*",
  "./*": "./src/*"
},
```

The only two exceptions to these patterns are `core` and the `vrembem` package. In `vrembem`, we also include a `root` export to include all custom properties output. In `core`, we also include entires to all available core SCSS modules as well as a root import:

```js
"exports": {
    ".": {
      "import": "./index.js",
      "require": "./dist/index.umd.cjs",
      "sass": "./index.scss",
      "style": "./dist/index.css"
    },
    "./dev/*": "./dev/*",
    "./dist/*": "./dist/*",
    "./root": "./root.scss",
    "./config": "./src/scss/modules/config.scss",
    "./usage": "./src/scss/modules/usage.scss",
    "./css": "./src/scss/modules/css.scss",
    "./palette": "./src/scss/modules/palette.scss",
    "./theme": "./src/scss/modules/theme.scss",
    "./*": "./src/*"
  },
```

Because of these changes, the following package properties have been removed as they are either not standard or are superseded by the exports property:

- `"main"`: Replaced with `"." > "import"` and `"." > "require"`
- `"sass"`: Replaced with `"." > "sass"`
- `"style"`: Replaced with `"." > "style"`