Allow passing maps to config.set mixin (#1917)
This PR updates the logic in `config.set` to now allow passing maps for more easily updating the config options map. This behavior is mirrored to how `css.set` works where the following parameter type are valid:
- `"key", "value"`
- `"key", ("key": "value")`
- `("key": "value"), null`
- `("key": ("key": "value")), null`

In cases where nested maps are provided, the root key is prefixed to subsequent keys. This allows the following:

```
@include config.set("asdf", (
  "one": 1,
  "two": 2
));

// Resulting options map:
$_options: (
  "asdf-one": 1,
  "asdf-two": 2
);
```