Add "form-control" and "transition" modules back in the "core" css module (#1925)
I've noticed that having a separate module for "form-control" and "transition" was a bit annoying and hard to remember so I moved all of those properties back under the "core" module. When using `css.get`, core is used as the default module so it's not even needed in the css reference calls.