Refactor theme module as an extension of the css module (#1928)
To help reduce complexity and improve usability of the theme module, this PR refactors it to be an extension of the CSS module instead of maintaining it's own custom property map. Internally, that means `theme` no longer maintains it's own custom properties map, but instead it stores theme properties in `css.$_variables`. To allow this, `css` module has been made more flexible by allowing `css.get` and `css.set` to more directly access the contents of `css.$_variables` map.

**Deprecated theme API**
- `theme.get()` replaced with `css.get()`
- `theme.log()` replaced with `css.log()`
- `theme.$_themes` usage replaced with `css.$_variables`

**Additional changes**
- Fixed issue with broken watch npm scripts.
- Fixed issue with section component not properly getting root style padding.