Add new themeStore core module for managing theme state (#1951)
This PR adds the new `themeStore` core module for managing and maintaining theme state. It provides an API for setting a theme, adding and removing possible themes as well as callbacks for when the store is first initialized and when a theme changes. Theme state is persisted via local storage and theme classes are applied to the `html` element.

**Additional changes**
- cssVar now takes an options param and allows setting a fallback value for when a CSS custom property is not found.
- Adds `prefix-themes` to core css custom properties.
- Applies new markup and logic to the `vrembem` package's vite demo including a theme toggle and `VITE_PACKAGE` variable.