Add css.defer function to better manage dependent custom property values and output (#1962)
This PR adds a new function `css.defer` to the CSS module that will prevent outputting custom properties if they're referenced in another custom property that isn't used. Once the dependent property is used, both custom properties will be output. This is done by setting a special value in the variables map which is evaluated in both `get` and `output` calls.

This PR also adds a necessary `form-control-size` mixin that handles setting the size custom property of form controls which relies on having available the values of `form-control-line-height` and `form-control-padding-y`. Since these values are not deferred, we use a mixin to set this property before it's used.