Refactor how form-controls handle size calculations (#1964)
This PR tries to simplify form-control sizing in a few ways. First, it removes the need to subtract border-width from padding. This ended up being an unnecessary calculation and resulted in extra complexity when trying to align form-control sizing. We also remove the need to set `min-width` and `min-height` on form controls. Instead, sizing is determined by the inherit properties set on the control such as line-height, padding and border-width.

Lastly, form-control components that do need to have their sizes explicitly set (such as checkbox, radio and switch) now do so using the refactored `form-control` library module. Width and height are calculated using line-height, padding and border-width properties and can be overridden on a component bases using custom properties.

**Additional changes**
- `minus-border-width` has been removed as we no longer subtract border width from padding values.
- `size()` has been refactored to allow setting a property prefix of min/max.
- `form-control-get-size()` function added to calculate and return form control size.
- `form-control-set-size()` mixin added to set the form-control-size custom property when it's needed.
- `form-control-size()` mixin added for setting the width/height properties using the form-control-size custom property. Can optionally set min/max width/height using the prefix parameter.