Improve support for multiple popovers on a single trigger (#1990)
This PR improves on the ability for popover triggers to have multiple popovers such as a menu dropdown and a tooltip. This is most commonly used with button icons where additional context may be needed for what the button does.

To accomplish this, the `isTooltip` property has been refactored to more accurately determine if a popover is indeed a tooltip. Tooltips now have additional guards set for when used together with click toggled popovers so that they no longer overlap and are handled more elegantly. For example:

- Tooltips can no longer be opened if a click popover is already open.
- Tooltips are displayed on `:focus-visible` instead of just `:focus`.
- The Role attribute is automatically applied if a popover has the tooltip modifier.
- Clicking a popover trigger will dismiss tooltips.
- Automatically applies the hover event to tooltip popovers.

**Additional Changes**
- Used a consistent arrow size value for all popovers.
- Improved error message for when a popover element is not found.
- Updates Vite demos to use package data. This allows for applying Vite demos in a more flexible way across multiple packages (future update).
- Refactors the FocusTrap module to copy code from focusable-selectors package and incorporate changes to prevent test errors.
- Improved Popover astro component to better allow for popover and tooltips on a single trigger and applies tooltips to button icons where needed.
- Adds documentation for using multiple popovers on a single trigger.