Add basic transition utilities (#1991)
This PR removes the temporary transition class in the popover component and adds two basic transition utilities to the utility module. This also sets up the utility component to run the build directory for output files. This will be more useful later when the utility component has been given custom properties support.