Refactor drawer and add custom properties support (#2008)
This PR primarily applies custom properties support to the drawer component. This is done in a similar way to how popover custom properties are applied in that they are provided fallback values and set the custom property names to override but do not output the custom properties directly. This PR also applies the new BEM naming function and removes redundant variables.