Fix code block copy button popover (#2019)
With recent changes to popover tooltip styles, this PR addresses the copy button popover which displays the "Copied!" text. Since this text should only appear on click, we make it a normal popover instead of a tooltip. We then pass the appropriate styles as css overrides and also adjust the spacing of code example copy buttons to better align in the toolbar area.