Migrate from using Popper JS to Floating UI (#2030)
This PR refactors our popover component to implement Floating UI and removes the use of Popper JS. This is done by importing the equivalent features and functionality that we were using with Popper and providing options objects as needed. We also rename the use of `overflow` to `shift` for the same reason floating UI updates this concept.