Fix hover event on non-tooltip popovers by tracking hover state (#2032)
This PR fixes the hover event on non-tooltip popovers by preventing the popover from closing when the trigger loses hover state. This is done by tracking the hover state on both the popover element and the trigger.

We're doing this using a new getter/setter on collection entries called `isHovered` which takes the event object as a parameter. This is then used to test if the event type is one of the mouse events for the hover state ("mouseenter" or "mouseleave") and if the event target matches either the popover trigger or popover element.

The mouse enter and leave handlers pass the event object to the setter and the mouse leave handler uses the getter to check if the popover should be closed or not.