Use inherited properties when creating collection entries (#2036)
This PR refactors how entries are created in components that extend the Collection class. Collection now provides a `createEntry` method that returns an entry object with inherited methods and copied properties to help reduce the amount of boilerplate setup needed in component entry registration. This PR also moves the teleport feature into the collections class so that all extending components now have that feature built in.

**Additional changes**
- The `data-[module]-config` data attribute has been simplified to `data-config` for all components that use this feature.
- Core has new `getElement` class for returning on HTML element using an ID. Can also be provided an HTML element directly. This allows functions to be provided either an ID string or HTML element.
- Module name is now a property on the created components that extend Collection.
- Base component settings are now provided via super param.
- Modal `required` property has been renamed to `isRequired`.
- Default values are no longer stored on component objects.