Refactor which custom properties are queried for in collection entries (#2051)
This PR refactors which custom properties are queried when `getCustomProps` is called. Instead of searching every key in the `entry.context.settings` object, we now use the `entry.getSetting("customProps")` method to return an array of all the props to query custom properties for.

As part of this update, `getSetting` has also been refactored into a modular helper function which is then imported and applied as the entry method `getSetting`. It can now take an options object for setting a fallback value in cases where a setting doesn't exist as well as a props array with the order of properties to check for the setting. This allows for adjusting the order or adding on to the settings query.