Create a plugin system for collections and entries (#2056)
This PR introduces the new plugin system for collections and entries. It allows for plugins to be created and registered to a collection and ties into the lifecycle of collections and entries via lifecycle hooks. The main requirement of a plugin is that it has a unique name, otherwise it's settings, methods and hooks that are used are up to the plugin author. As part of this system, the teleport module has been converted into the first plugin along with a debug plugin for better understanding when and in what order hooks are fired.

**Additional Changes**
- Modal and Drawer Vite demo pages have been updated to the new format.
- Teleport plugin has been applied to the docs website.
- Core JS modules have been reorganized into helper and utility directories along with the new plugins directory where all future plugins will be stored.
- Manifest index export files have been added to all sub JS directories to help keep imports/exports more organized.
- Entry `before/afterMount` and `before/afterUnmount` have been replaced with  the more simple `onMount` and `onUnmount` hooks.
- Added new `processOptions()` method to collections to allow passing plugins as part of the options object but removing it before options are merged into settings.
- Created new `pluginsArray` class which extends the base `Array` class. It creates the object used for managing plugins in a collection.
- Renamed `lifecycleHook` to `maybeRunMethod` for clarity.
- Renamed `teleport` utility to `teleportElement` for clarity.