Create new propStore plugin (#2081)
## What changed?

This PR creates the `propStore` plugin which allows the creation of a reactive property on entry objects that can also be stored in local storage. This allows for creating reactive functionality based on the state of an entry as well as persisting the state across sessions. The local storage object can be a accessed through the plugin instance (`plugin.store.[get||set]`) or the entry object (`entry.store`).

With this update, all plugins now get processed through `createPluginObject` which allows subscribing to component lifecycle hooks as well as setting a custom plugin name value for multiple uses in the same component.

### Use cases

This plugin can be used in many ways, but here are some practical use cases in the context of drawer and modal components. 

**Persisting Drawer State**

In this example we want to persist the open state of an inline drawer across sessions. To do this, we setup a watch on the `inlineState` property of entries with the default value of the local storage. We update the value of local storage using the `condition` function which only stores the value if the drawer is opened or closed. We use the `onChange` method to apply the new state whenever it changes.

```js
import Drawer from "@vrembem/drawer";
import { propStore } from "@vrembem/core";

const drawer = new Drawer({
  plugins: [
    propStore({
      prop: "inlineState",
      value: (entry) => entry.store,
      condition: (entry) => ["opened", "closed"].includes(entry.state),
      onChange: (entry) => entry.applyState()
    })
  ]
});
```

**Persisting Modal State**

In this example we want to persist the open state of modals across sessions. To do this, we watch the `state` (by default) property on entry objects. We use the `onChange` callback to store the stack array in local storage. We then setup a `afterMount` lifecycle hook to open all modals in the stack array after the component has mounted.

```js
import Modal from "@vrembem/modal";
import { propStore } from "@vrembem/core";

const modal = new Modal({
  plugins: [
    propStore({
      onChange(entry) {
        const stackValue = entry.parent.stack.value.map((entry) => entry.id);
        this.store.set("stackValue", stackValue);
      },
      afterMount(parent) {
        for (const entry of this.store.get("stackValue", [])) {
          parent.open(entry, false);
        }
      }
    })
  ]
});
```

### Additional changes

As part of this update, the following changes have also been made:
- `localStorage` utility get method now allows setting a fallback value in cases where a value in local storage does not exist.
- Vite demo pages have been updated to implement `propStore` usage.
- Drawer no longer has built in state persistence functionality. This is now achieved using the `propStore` plugin.
- Drawer state has been refactored to more concisely determine the appropriate state using the new `applyState()` method. This is used internally on mount and when the drawer mode switches but has also been made public. This replaces the use of `applyInitialState()` and `applyInlineState()` which have both been removed.
- Drawer now using `setState(value)` to set the `drawer.state` property instead of the previously used getter/setter.
- Documentation website has been updated to implement feature parity with `propStore` usage and Drawer's previously built in state persistence feature.